+++ modules/Emails/include/ListView/ListViewDataEmails.php
@@ -44,119 +44,12 @@ if (!defined('sugarEntry') || !sugarEntry) {
 
 require_once('include/ListView/ListViewData.php');
 
-class SuiteException extends Exception {}
+include_once 'include/Exceptions/SuiteException.php';
+include_once 'modules/Emails/Folder.php';
+include_once 'modules/Emails/include/ListView/ListViewDataEmailsSearchOnCrm.php';
+include_once 'modules/Emails/include/ListView/ListViewDataEmailsSearchOnIMap.php';
 
-/**
- * Class Folder
- *
- * private model class for ListViewDataEmails::getListViewData()
- * represent a fake SugarBean:
- * in legacy logic, Folder ID equals to an Inbound Email ID
- */
-class Folder extends SugarBean
-{
 
-    /**
-     * private
-     * @var DBManager $db
-     */
-    public $db;
-
-    /**
-     * private, use Folder::getId() instead
-     * @var string UUID in folders table
-     */
-    public $id;
-
-    /**
-     * private, use Folder::getType() instead
-     * @var string folder type
-     */
-    protected $type;
-
-    /**
-     * Folder constructor.
-     */
-    public function __construct() {
-        parent::__construct();
-        $this->db = DBManagerFactory::getInstance();
-        $this->id = null;
-        $this->type = "inbound";
-    }
-
-    /**
-     * @param int|string $folderId - (should be string, int type is legacy)
-     * @param bool $encode (legacy, unused)
-     * @param bool $deleted (legacy, unused)
-     * @return null|string (folder ID)
-     * @throws SuiteException
-     */
-    public function retrieve($folderId = -1, $encode = true, $deleted = true)
-    {
-        if(isValidId($folderId)) {
-
-            // TODO: use parent::retrieve here
-            $result = $this->db->query('SELECT * FROM folders WHERE id="' . $folderId . '"');
-            $row = $this->db->fetchByAssoc($result);
-
-            // get the root of the tree
-            // is the id of the root node is the same as the inbound email id
-
-            if (empty($row['parent_folder'])) {
-
-                // root node (inbound)
-
-                $this->id = $row['id'];
-
-            } else {
-
-                // child node
-
-                $this->id = $row['parent_folder'];
-                $this->type = $row['folder_type'];
-
-            }
-
-        } else {
-            throw new SuiteException("Invalid or empty Email Folder ID");
-        }
-
-        return $this->id;
-    }
-
-    /**
-     * @param array $request
-     * @return Folder
-     * @throws SuiteException
-     */
-    public function retrieveFromRequest($request) {
-
-        if (isset($request['folders_id']) && !empty($request['folders_id'])) {
-
-            $foldersId = $request['folders_id'];
-            $this->retrieve($foldersId);
-        } else {
-            $GLOBALS['log']->warn("Empty or undefined Email Folder ID");
-        }
-
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getType() {
-        return $this->type;
-    }
-
-    /**
-     * @return null|string
-     */
-    public function getId() {
-        return $this->id;
-    }
-
-}
 
 class ListViewDataEmails extends ListViewData
 {
@@ -814,560 +707,7 @@ class ListViewDataEmails extends ListViewData
 
 }
 
-abstract class ListViewDataEmailsSearchAbstract {
 
-    /**
-     * @var ListViewDataEmails
-     */
-    protected $lvde;
 
-    public function __construct(ListViewDataEmails $listViewDataEmails) {
-        $this->lvde = $listViewDataEmails;
-    }
 
-}
 
-class ListViewDataEmailsSearchOnCrm extends ListViewDataEmailsSearchAbstract {
-
-    /**
-     * @param array $filterFields
-     * @param array $request $_REQUEST
-     * @param string $where
-     * @param InboundEmail $inboundEmail
-     * @param array $params
-     * @param Email $seed
-     * @param bool $singleSelect
-     * @param ??? $id
-     * @param int $limit
-     * @param User $currentUser
-     * @param string $idField
-     * @param int $offset
-     * @return array
-     */
-    public function search($filterFields, $request, $where, InboundEmail $inboundEmail, $params, Email $seed, $singleSelect, $id, $limit, User $currentUser, $idField, $offset) {
-        // Fix fields in filter fields
-
-        $filterFields = $this->lvde->fixFieldsInFilter($filterFields, $request, $where);
-
-
-        // Filter imported emails based on the UID of the results from the IMap server
-
-        if(!empty($where)) {
-            $where .= ' AND ';
-        }
-        $crmWhere = $where . 'mailbox_id LIKE ' .'"' . $inboundEmail->id . '"';
-
-
-        // Populates CRM fields
-
-        $crmQueryArray = $this->lvde->getCrmQueryArray($crmWhere, $filterFields, $params, $seed, $singleSelect);
-
-
-        // get crm emails query
-
-        $crmEmailsQuery = $this->lvde->getCrmEmailsQuery($crmQueryArray, $params);
-
-
-
-        $crmEmails = $this->lvde->db->query($crmEmailsQuery);
-
-        $this->lvde->setVariableName($seed->object_name, $crmWhere, $this->lvde->listviewName, $id);
-
-        SugarVCR::erase($seed->module_dir);
-        $this->lvde->seed =& $seed;
-        $totalCounted = empty($GLOBALS['sugar_config']['disable_count_query']);
-        $_SESSION['MAILMERGE_MODULE_FROM_LISTVIEW'] = $seed->module_dir;
-        if(empty($request['action']) || $request['action'] != 'Popup'){
-            $_SESSION['MAILMERGE_MODULE'] = $seed->module_dir;
-        }
-
-        $this->lvde->setVariableName($seed->object_name, $where, $this->lvde->listviewName, $id);
-
-        $this->lvde->seed->id = '[SELECT_ID_LIST]';
-
-        // if $params tell us to override all ordering
-        if(!empty($params['overrideOrder']) && !empty($params['orderBy'])) {
-            $order = $this->lvde->getOrderBy(strtolower($params['orderBy']), (empty($params['sortOrder']) ? '' : $params['sortOrder'])); // retreive from $_REQUEST
-        }
-        else {
-            $order = $this->lvde->getOrderBy(); // retreive from $_REQUEST
-        }
-
-        // still empty? try to use settings passed in $param
-        if(empty($order['orderBy']) && !empty($params['orderBy'])) {
-            $order['orderBy'] = $params['orderBy'];
-            $order['sortOrder'] =  (empty($params['sortOrder']) ? '' : $params['sortOrder']);
-        }
-
-        if (empty($params['skipOrderSave'])) { // don't save preferences if told so
-            $currentUser->setPreference('listviewOrder', $order, 0, $this->lvde->var_name); // save preference
-        }
-
-        $data = array();
-
-        $temp = clone $seed;
-
-        $rows = array();
-        $count = 0;
-        $idIndex = array();
-        $id_list = '';
-
-        while(($row = $this->lvde->db->fetchByAssoc($crmEmails)) != null)
-        {
-            if($count < $limit)
-            {
-                $id_list .= ',\''.$row[$idField].'\'';
-                $idIndex[$row[$idField]][] = count($rows);
-                $rows[] = $seed->convertRow($row);
-            }
-            $count++;
-        }
-
-        if (!empty($id_list))
-        {
-            $id_list = '('.substr($id_list, 1).')';
-        }
-
-        SugarVCR::store($this->lvde->seed->module_dir,  $crmEmailsQuery);
-        if($count != 0) {
-            //NOW HANDLE SECONDARY QUERIES
-            if(!empty($ret_array['secondary_select'])) {
-                $secondary_query = $ret_array['secondary_select'] . $ret_array['secondary_from'] . ' WHERE '.$this->lvde->seed->table_name.'.id IN ' .$id_list;
-                if(isset($ret_array['order_by']))
-                {
-                    $secondary_query .= ' ' . $ret_array['order_by'];
-                }
-
-                $secondary_result = $this->lvde->db->query($secondary_query);
-
-                $ref_id_count = array();
-                while($row = $this->lvde->db->fetchByAssoc($secondary_result)) {
-
-                    $ref_id_count[$row['ref_id']][] = true;
-                    foreach($row as $name=>$value) {
-                        //add it to every row with the given id
-                        foreach($idIndex[$row['ref_id']] as $index){
-                            $rows[$index][$name]=$value;
-                        }
-                    }
-                }
-
-                $rows_keys = array_keys($rows);
-                foreach($rows_keys as $key)
-                {
-                    $rows[$key]['secondary_select_count'] = count($ref_id_count[$rows[$key]['ref_id']]);
-                }
-            }
-
-            // retrieve parent names
-            if(!empty($filter_fields['parent_name']) && !empty($filter_fields['parent_id']) && !empty($filter_fields['parent_type'])) {
-                foreach($idIndex as $id => $rowIndex) {
-                    if(!isset($post_retrieve[$rows[$rowIndex[0]]['parent_type']])) {
-                        $post_retrieve[$rows[$rowIndex[0]]['parent_type']] = array();
-                    }
-                    if(!empty($rows[$rowIndex[0]]['parent_id'])) $post_retrieve[$rows[$rowIndex[0]]['parent_type']][] = array('child_id' => $id , 'parent_id'=> $rows[$rowIndex[0]]['parent_id'], 'parent_type' => $rows[$rowIndex[0]]['parent_type'], 'type' => 'parent');
-                }
-                if(isset($post_retrieve)) {
-                    $parent_fields = $seed->retrieve_parent_fields($post_retrieve);
-                    foreach($parent_fields as $child_id => $parent_data) {
-                        //add it to every row with the given id
-                        foreach($idIndex[$child_id] as $index){
-                            $rows[$index]['parent_name']= $parent_data['parent_name'];
-                        }
-                    }
-                }
-            }
-
-            $pageData = array();
-
-            reset($rows);
-            while($row = current($rows)){
-
-                $temp = clone $seed;
-                $dataIndex = count($data);
-
-                $temp->setupCustomFields($temp->module_dir);
-                $temp->loadFromRow($row);
-                if (empty($this->lvde->seed->assigned_user_id) && !empty($temp->assigned_user_id)) {
-                    $this->lvde->seed->assigned_user_id = $temp->assigned_user_id;
-                }
-                if($idIndex[$row[$idField]][0] == $dataIndex){
-                    $pageData['tag'][$dataIndex] = $temp->listviewACLHelper();
-                }else{
-                    $pageData['tag'][$dataIndex] = $pageData['tag'][$idIndex[$row[$idField]][0]];
-                }
-                $data[$dataIndex] = $temp->get_list_view_data();
-                $detailViewAccess = $temp->ACLAccess('DetailView');
-                $editViewAccess = $temp->ACLAccess('EditView');
-                $pageData['rowAccess'][$dataIndex] = array('view' => $detailViewAccess, 'edit' => $editViewAccess);
-                $additionalDetailsAllow = $this->lvde->additionalDetails && $detailViewAccess && (file_exists(
-                            'modules/' . $temp->module_dir . '/metadata/additionalDetails.php'
-                        ) || file_exists('custom/modules/' . $temp->module_dir . '/metadata/additionalDetails.php'));
-                $additionalDetailsEdit = $editViewAccess;
-                if($additionalDetailsAllow) {
-                    if($this->lvde->additionalDetailsAjax) {
-                        $ar = $this->lvde->getAdditionalDetailsAjax($data[$dataIndex]['ID']);
-                    }
-                    else {
-                        $additionalDetailsFile = 'modules/' . $this->lvde->seed->module_dir . '/metadata/additionalDetails.php';
-                        if(file_exists('custom/modules/' . $this->lvde->seed->module_dir . '/metadata/additionalDetails.php')){
-                            $additionalDetailsFile = 'custom/modules/' . $this->lvde->seed->module_dir . '/metadata/additionalDetails.php';
-                        }
-                        require_once($additionalDetailsFile);
-                        $ar = $this->lvde->getAdditionalDetails($data[$dataIndex],
-                            (empty($this->lvde->additionalDetailsFunction) ? 'additionalDetails' : $this->lvde->additionalDetailsFunction) . $this->lvde->seed->object_name,
-                            $additionalDetailsEdit);
-                    }
-                    $pageData['additionalDetails'][$dataIndex] = $ar['string'];
-                    $pageData['additionalDetails']['fieldToAddTo'] = $ar['fieldToAddTo'];
-                }
-                next($rows);
-            }
-        }
-        $nextOffset = -1;
-        $prevOffset = -1;
-        $endOffset = -1;
-        if($count > $limit) {
-            $nextOffset = $offset + $limit;
-        }
-
-        if($offset > 0) {
-            $prevOffset = $offset - $limit;
-            if($prevOffset < 0)$prevOffset = 0;
-        }
-        $totalCount = $count + $offset;
-
-        if( $count >= $limit && $totalCounted){
-            $totalCount  = $this->lvde->getTotalCount($crmEmailsQuery);
-        }
-        SugarVCR::recordIDs($this->lvde->seed->module_dir, array_keys($idIndex), $offset, $totalCount);
-        $module_names = array(
-            'Prospects' => 'Targets'
-        );
-        $endOffset = (floor(($totalCount - 1) / $limit)) * $limit;
-        $pageData['ordering'] = $order;
-        $pageData['ordering']['sortOrder'] = $this->lvde->getReverseSortOrder($pageData['ordering']['sortOrder']);
-        //get url parameters as an array
-        $pageData['queries'] = $this->lvde->callGenerateQueries($pageData['ordering']['sortOrder'], $offset, $prevOffset, $nextOffset,  $endOffset, $totalCounted);
-        //join url parameters from array to a string
-        $pageData['urls'] = $this->lvde->callGenerateURLS($pageData['queries']);
-        $pageData['offsets'] = array( 'current'=>$offset, 'next'=>$nextOffset, 'prev'=>$prevOffset, 'end'=>$endOffset, 'total'=>$totalCount, 'totalCounted'=>$totalCounted);
-        $pageData['bean'] = array('objectName' => $seed->object_name, 'moduleDir' => $seed->module_dir, 'moduleName' => strtr($seed->module_dir, $module_names));
-        $pageData['stamp'] = $this->lvde->stamp;
-        $pageData['access'] = array('view' => $this->lvde->seed->ACLAccess('DetailView'), 'edit' => $this->lvde->seed->ACLAccess('EditView'));
-        $pageData['idIndex'] = $idIndex;
-        if(!$this->lvde->seed->ACLAccess('ListView')) {
-            $pageData['error'] = 'ACL restricted access';
-        }
-
-        $queryString = '';
-
-        if( isset($request["searchFormTab"]) && $request["searchFormTab"] == "advanced_search" ||
-            isset($request["type_basic"]) && (count($request["type_basic"] > 1) || $request["type_basic"][0] != "") ||
-            isset($request["module"]) && $request["module"] == "MergeRecords")
-        {
-            $queryString = "-advanced_search";
-        }
-        else if (isset($request["searchFormTab"]) && $request["searchFormTab"] == "basic_search")
-        {
-            // TODO: figure out what was the SearchFormReports???
-            if($seed->module_dir == "Reports") $searchMetaData = SearchFormReports::retrieveReportsSearchDefs();
-            else $searchMetaData = SearchForm::retrieveSearchDefs($seed->module_dir); // TODO: figure out which SearchForm is it?
-
-            $basicSearchFields = array();
-
-            if( isset($searchMetaData['searchdefs']) && isset($searchMetaData['searchdefs'][$seed->module_dir]['layout']['basic_search']) )
-                $basicSearchFields = $searchMetaData['searchdefs'][$seed->module_dir]['layout']['basic_search'];
-
-            foreach( $basicSearchFields as $basicSearchField)
-            {
-                $field_name = (is_array($basicSearchField) && isset($basicSearchField['name'])) ? $basicSearchField['name'] : $basicSearchField;
-                $field_name .= "_basic";
-                if( isset($request[$field_name])  && ( !is_array($basicSearchField) || !isset($basicSearchField['type']) || $basicSearchField['type'] == 'text' || $basicSearchField['type'] == 'name') )
-                {
-                    // Ensure the encoding is UTF-8
-                    $queryString = htmlentities($request[$field_name], null, 'UTF-8');
-                    break;
-                }
-            }
-        }
-
-        $ret =  array('data'=>$data , 'pageData'=>$pageData, 'query' => $queryString);
-
-        return $ret;
-    }
-
-}
-
-class ListViewDataEmailsSearchOnIMap extends ListViewDataEmailsSearchAbstract {
-
-
-
-    /**
-     * @param Email $seed
-     * @param string $where
-     * @param string $id
-     * @param InboundEmail $inboundEmail
-     * @param array $filter
-     * @param Folder $folderObj
-     * @param User $currentUser
-     * @param string $folder
-     * @param int $limit
-     * @param ??? $limitPerPage
-     * @return array
-     */
-    public function search(Email $seed, $where, $id, InboundEmail $inboundEmail, $filter, Folder $folderObj, User $currentUser, $folder, $limit, $limitPerPage) {
-
-
-        // Create the data structure which are required to view a list view.
-        require_once 'include/SearchForm/SearchForm2.php';
-        $this->lvde->seed =& $seed;
-        $this->lvde->setVariableName($seed->object_name, $where, $this->lvde->listviewName, $id);
-        $this->lvde->seed->id = '[SELECT_ID_LIST]';
-
-        if (!empty($params['overrideOrder']) && !empty($params['orderBy'])) {
-            $order = $this->lvde->getOrderBy(
-                strtolower($params['orderBy']),
-                (empty($params['sortOrder']) ? '' : $params['sortOrder'])
-            );
-        } else {
-            $order = $this->lvde->getOrderBy();
-        }
-
-        // still empty? try to use settings passed in $param
-        if (empty($order['orderBy']) && !empty($params['orderBy'])) {
-            $order['orderBy'] = $params['orderBy'];
-            $order['sortOrder'] = (empty($params['sortOrder']) ? '' : $params['sortOrder']);
-        }
-
-        // TODO: figure out why was it for?
-        $orderby = $order['orderBy'];
-        if (strpos($order['orderBy'], '.') && ($order['orderBy'] != "report_cache.date_modified")) {
-            $orderby = substr($order['orderBy'], strpos($order['orderBy'], '.') + 1);
-        }
-
-
-        $page = 0;
-        $offset = 0;
-        if (isset($request['Emails2_EMAIL_offset'])) {
-            if ($request['Emails2_EMAIL_offset'] !== "end") {
-                $offset = $request['Emails2_EMAIL_offset'];
-                $page = $offset / $limitPerPage;
-            }
-        }
-
-        // Get emails from email server
-        $emailServerEmails = $inboundEmail->checkWithPagination($offset, $limitPerPage, $order, $filter);
-
-        $total = $emailServerEmails['mailbox_info']['Nmsgs']; // + count($importedEmails['data']);
-        if ($page === "end") {
-            $offset = $total - $limitPerPage;
-        }
-
-
-        /// Populate the data and its fields from the email server
-        $request['uids'] = array();
-
-        foreach ($emailServerEmails['data'] as $h => $emailHeader) {
-
-
-            $emailRecord = $this->lvde->getEmailRecord($folderObj, $emailHeader, $seed, $inboundEmail, $currentUser, $folder);
-            if($emailRecord === false) {
-                continue;
-            }
-
-            $data[] = $emailRecord;
-            $pageData['rowAccess'][$h] = array('edit' => true, 'view' => true);
-            $pageData['additionalDetails'][$h] = '';
-            $pageData['tag'][$h]['MAIN'] = 'a';
-        }
-
-
-        // Filter imported emails based on the UID of the results from the IMap server
-        $crmWhere = $where . ' AND mailbox_id LIKE ' . '"' . $inboundEmail->id . '"';
-
-        $ret_array['inner_join'] = '';
-        if (!empty($this->lvde->seed->listview_inner_join)) {
-            $ret_array['inner_join'] = ' ' . implode(' ', $this->lvde->seed->listview_inner_join) . ' ';
-        }
-
-        if (!is_array($params)) {
-            $params = array();
-        }
-        if (!isset($params['custom_select'])) {
-            $params['custom_select'] = '';
-        }
-        if (!isset($params['custom_from'])) {
-            $params['custom_from'] = '';
-        }
-        if (!isset($params['custom_where'])) {
-            $params['custom_where'] = '';
-        }
-        if (!isset($params['custom_order_by'])) {
-            $params['custom_order_by'] = '';
-        }
-
-        $this->lvde->setVariableName($seed->object_name, $crmWhere, $this->lvde->listviewName, $id);
-
-        $this->lvde->seed->id = '[SELECT_ID_LIST]';
-
-        $nextOffset = -1;
-        $prevOffset = -1;
-        $endOffset = 0;
-
-
-        if ($total > $limitPerPage) {
-            $nextOffset = $offset + $limitPerPage;
-        }
-
-        if ($nextOffset >= $total) {
-            $nextOffset = $total;
-        }
-
-        if ($page > 0) {
-            $prevOffset = $offset - $limitPerPage;
-            if ($prevOffset < 0) {
-                $prevOffset = 0;
-            }
-        }
-
-        if ($total < $limitPerPage) {
-            $prevOffset = -1;
-            $nextOffset = -1;
-        }
-
-        if ($total > 0) {
-            $endOffset = $total / $limitPerPage;
-        }
-
-        $pageData['offsets']['current'] = $offset;
-        $pageData['offsets']['total'] = $total;
-        $pageData['offsets']['next'] = $nextOffset;
-        $pageData['offsets']['prev'] = $prevOffset;
-        $pageData['offsets']['end'] = ceil($endOffset);
-
-        $queries = array('baseUrl', 'endPage', 'nextPage', 'orderBy');
-
-        if ((int)$pageData['offsets']['current'] >= $limitPerPage) {
-            $queries[] = 'prevPage';
-            $queries[] = 'startPage';
-        }
-
-        foreach ($queries as $query) {
-
-            if ($total < $limitPerPage || $nextOffset >= $total) {
-                if (isset($pageData['queries'][$query])) {
-                    unset($pageData['queries'][$query]);
-                }
-            } else {
-                $pageData['queries'][$query]['module'] = "Emails";
-                $pageData['queries'][$query]['action'] = "index";
-                $pageData['queries'][$query]['parentTab'] = "Activities";
-                $pageData['queries'][$query]['ajax_load'] = "0";
-                $pageData['queries'][$query]['loadLanguageJS'] = "1";
-                $pageData['queries'][$query]['searchFormTab'] = "advanced_search";
-                $pageData['queries'][$query]['lvso'] = "DESC";
-
-                $pageData['urls'][$query] = 'index.php?module=Emails&action=index&parentTab=Activities&searchFormTab=advanced_search&query=true&current_user_only_basic=0&button=Search&lvso=DESC';
-
-            }
-        }
-
-        // TODO: UNDEFINED: refactor current_query_by_page in the list view
-        // inject post values
-        $request['folder'] = $folder;
-        $request['folder_type'] = $folderObj->getType();
-        $request['inbound_email_record'] = $inboundEmail->id;
-
-
-        // TODO: TASK: UNDEFINED - HANDLE in second filter after IMap
-        $endOffset = floor(($total - 1) / $limit) * $limit;
-        $pageData['queries'] = $this->lvde->callGenerateQueries(
-            $pageData['ordering']['sortOrder'],
-            $offset,
-            $prevOffset,
-            $nextOffset,
-            $endOffset,
-            $total
-        );
-        $pageData['offsets'] = array(
-            'current' => $offset,
-            'next' => $nextOffset,
-            'prev' => $prevOffset,
-            'end' => $endOffset,
-            'total' => $total,
-            'totalCounted' => $total
-        );
-
-        $pageData['ordering'] = $order;
-        $pageData['ordering']['sortOrder'] = $this->lvde->getReverseSortOrder($pageData['ordering']['sortOrder']);
-        //get url parameters as an array
-        //join url parameters from array to a string
-        $pageData['urls'] = $this->lvde->callGenerateURLS($pageData['queries']);
-        $module_names = array(
-            'Prospects' => 'Targets'
-        );
-        $pageData['bean'] = array(
-            'objectName' => $seed->object_name,
-            'moduleDir' => $seed->module_dir,
-            'moduleName' => strtr($seed->module_dir, $module_names)
-        );
-        $pageData['stamp'] = $this->lvde->stamp;
-        $pageData['access'] = array(
-            'view' => $this->lvde->seed->ACLAccess('DetailView'),
-            'edit' => $this->lvde->seed->ACLAccess('EditView')
-        );
-        if (!$this->lvde->seed->ACLAccess('ListView')) {
-            $pageData['error'] = 'ACL restricted access';
-        }
-
-
-        if ( $this->lvde->isRequestedSearchAdvanced($request) ) {
-            $queryString = "-advanced_search";
-        } else {
-            if ( $this->lvde->isRequestedSearchBasic($request) ) {
-
-                // SearchForm is a (SearchForm2)
-                $searchMetaData = SearchForm::retrieveSearchDefs($seed->module_dir);
-
-                $basicSearchFields = array();
-
-                if (
-                    isset($searchMetaData['searchdefs']) &&
-                    isset($searchMetaData['searchdefs'][$seed->module_dir]['layout']['basic_search'])
-                ) {
-                    $basicSearchFields = $searchMetaData['searchdefs'][$seed->module_dir]['layout']['basic_search'];
-                }
-
-                foreach ($basicSearchFields as $basicSearchField) {
-                    $field_name = (is_array($basicSearchField) && isset($basicSearchField['name']))
-                        ? $basicSearchField['name'] : $basicSearchField;
-                    $field_name .= "_basic";
-                    if (
-                        isset($request[$field_name]) &&
-                        (
-                            !is_array($basicSearchField) ||
-                            !isset($basicSearchField['type']) ||
-                            $basicSearchField['type'] == 'text' ||
-                            $basicSearchField['type'] == 'name'
-                        )
-                    ) {
-                        // Ensure the encoding is UTF-8
-                        $queryString = htmlentities($request[$field_name], null, 'UTF-8');
-                        break;
-                    }
-                }
-            } else {
-                $GLOBALS['log']->warn("Unknown requested search type");
-            }
-        }
-
-        $request['email_uids'] = $this->lvde->getEmailUIds($data);
-
-
-        $ret = array('data' => $data, 'pageData' => $pageData, 'query' => $queryString);
-
-        return $ret;
-    }
-
-}
